Fixed float64 warning

+++ lmu/lmu.py
@@ -1005,7 +1005,7 @@ class FFTLayer(Layer):
         )
 
         impulse = tf.reshape(
-            tf.eye(self.seq_length, 1, dtype=tf.dtypes.float64), (1, self.seq_length, 1)
+            tf.eye(self.seq_length, 1), (1, self.seq_length, 1)
         )
 
         self.impulse_response = tf.squeeze(tf.transpose(delay_layer(impulse)), [-1])
